Integrate Docker Compose with Aktualizr
When an docker-compose file update is received via a secondary, run
docker-compose commands to pull and start the new containers.

The process occurs in the background without the need to reboot the device.

If an error occurs during the process, the update is rolled back to the
previous docker-compose file.

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>
Signed-off-by: Jeremias Cordoba <jeremias.cordoba@toradex.com>