Use non-deprecated version of boost::bind
Boost 1.76.0 gives the following warning with the old import:

"The practice of declaring the Bind placeholders (_1, _2, ...) in the global
namespace is deprecated. Please use <boost/bind/bind.hpp> + using namespace
boost::placeholders, or define BOOST_BIND_GLOBAL_PLACEHOLDERS to retain the
current behavior."

We are not using the _1, _2, etc placeholders, so this is a simple change.

Signed-off-by: Phil Wise <phil@phil-wise.com>