Rework OSTreeObject to store a OSTreeHash rather than string object_name_
OSTreeObject used std::string object_name_ to store its own identity. This was
handy, but that string was basically a serialization of OSTreeHash and
OstreeObjectType. The bulk of this change is about moving away from the
'string' representation to a pair of OSTreeHash hash_ and OstreeObjectType
type_.

It also adds a OSTreeObject::Fsck() operation, which can be used to check for
corrupted objects. Adding this to garage-push will come as a separate commit.

Signed-off-by: Phil Wise <phil@phil-wise.com>