Use default credentials.zip to fix more tests
These tests only need a credentials.zip that looks sane: it doesn't need to
actually work. Use the 'tests/test_data/credentials.zip' to provide it. There
is one test for garage-deploy that needs a credentials.zip that authenticates
in order to get to the point where it triggers the error message we are looking
for. Move that one to be conditional on having SOTA_PACKED_CREDENTIALS.

Signed-off-by: Phil Wise <phil@phil-wise.com>