Set custom hardware information via an explicit call
Previously, custom hardware information was set by an optional parameter on the
update calls, e.g. Aktualizr::RunForever(). This added a bunch of complexity
around the core update loop that isn't really necessary. The only user of this
at the moment sets it to a static value on the command line.

Move this to Aktualizr::SetCustomHardwareInfo(Json::Value)

Signed-off-by: Phil Wise <phil@phil-wise.com>