Report correct OSTree hash when currently booted target is absent
This is a fix for https://github.com/uptane/aktualizr/issues/1 "Aktualizr
reports incorrect OSTree hash if it doesn't find the currently-booting version
in its targets".

When we can't find the original Uptane Target that matches the currently booted
hash, return one that has the correct hash. This means that the manifest
returned to the server when a factory-fresh image is booted has enough
information to allow the server to figure out what is running. Before this
change the Uptane::Target::Unknown() was reported, which has a fixed hash of
the empty string.

There are a couple of tests, one at the OstreeManager level, and a second that
checks the output of SotaUptaneClient::AssembleManifest().

Signed-off-by: Phil Wise <phil@phil-wise.com>