Fix flakey "test_aktualizr_kill" test
There wasn't anything reading the stdout of aktualizr when the output was set
to capture. As long as 'not much' output was written by aktualizr then it
didn't block and the test would get to the point where it can finish. There
doesn't seem to be an easy way to handle this in Python, so I wrote a quick
utility to copy the stdout to an in-memory buffer in a background thread.

Signed-off-by: Phil Wise <phil@phil-wise.com>