Check return value of recv(). Fix test_ip_secondary_rotation
The return value of recv() wasn't being checked before passing it to
DequeueBuffer::HaveEnqueued(). The negative value was being cast to an unsigned
size_t, which was causing DequeueBuffer::HaveEnqueued() to abort with an
assert because it thought there was an overflow.

With this change "ctest --repeat-until-fail 10 --output-on-failure -R
test_ip_secondary_rotation" passes on my machine.

Signed-off-by: Phil Wise <phil@phil-wise.com>