Don't fail metadata check if IP Secondaries don't support Root version requests.
Older versions don't support it, so if you run a newer Primary expecting
the support with an older Secondary that doesn't, the metadata will fail
to get sent no matter what you do. (Well, you can remove the Secondary,
but that's pretty nasty.) This change makes the check more tolerant. If
there's a problem with verifying the Root metadata (especially if an
intermediate Root was skipped), the verification will still fail, but
now at a slightly later phase (which is where it would've failed
before).

Signed-off-by: Patti Vacek <pattivacek@gmail.com>