Delay parts of provisioning that require online access
This is a significant reworking of the provisioning process to move the parts
that require access to the server as late as possible.  The goal here is
supporting updates delivered by physical media (i.e. offline).

I've made PrimaryEcuSerial() PrimaryHardwareIdentifier() and DeviceId() lazily
initialized, so they are set up on first use, either from the local storage or
by creating them. This removes some of the implicit dependencies between stages
of initialization.

The big win is that finalizeAfterReboot() is called before
Provisioner::Attempt(), which opens the door to only requiring provisioning
when on-line updates are attempted. This isn't the case quite yet, because
SotaUptaneClient::initialize() won't complete without online provisioning. Lots
of the tests work on this assumption, so some reworking of those will be
required in the future too.

Signed-off-by: Phil Wise <phil@phil-wise.com>