torizon: Fix manifest not containing installation results after certain metadata failures
In some cases where the metadata verification failed an installation result was not logged
in the manifest. This caused issues where an error in the client was not translated as an
update failure back to the OTA server.

Fix this by making sure the manifest contains an installation result in these cases.

Related-to: TOR-2249

Signed-off-by: Jeremias Cordoba <jeremias.cordoba@toradex.com>