Fix build breakage caused by git CVE-2022-24765 fixes
The Docker-ness of the CI means that the git checkout and build are done with
different users. This causes git releases with the fix for the above CVE to
abort with an error "fatal: unsafe repository ... is owned by someone else"

In order to fix this we explicitly mark the git repo as 'safe' during the build
process.

Signed-off-by: Phil Wise <phil@phil-wise.com>