Only write ManagedSecondary keys once
Previously this was overwriting the public and private keys of the managed
secondary on every startup. This works as long at the power doesn't go off at
the wrong time. We've seen examples where the file ends up empty after a power
off and/or the disk filling up.

This sequence of operations appears to just be a bug. Change the logic to only
write the keys once, and add a sync() call for good measure.

Signed-off-by: Phil Wise <phil@phil-wise.com>