Manually fetch the submodules in github CI.
The automatic fetching logic breaks in really confusing ways when the
submodule origin has changed. It really doesn't make sense. This is a
really naïve workaround. It is hopefully just a temporary fix that can
be reverted at a later time, but at least for now, it seems to be
necessary in order to get CI to pass on this branch.

Signed-off-by: Patti Vacek <pattivacek@gmail.com>