Limit event number sent in a single request
Allow limiting a number of events to be send in a scope of a single
request. This helps to avoid sending a big body if events have got
accumulated in the DB somehow. It's needed if a server, receiving events,
has a size limit of request body it can digest.

Signed-off-by: Mike <mike.sul@foundries.io>