Log installation failure when failing to update/verify metadata
When updating and verifying the metadata, there are errors where an
installation failure is not stored. Without a failure the server will
continue to send the same "bad" metadata causing the overall system
to be stuck.

With this we store an installation failure if the error requires new
metadata to be sent by the server to recover.

Related-to: TOR-2780

Signed-off-by: Jeremias Cordoba <jeremias.cordoba@toradex.com>