dockertarballloader: Handle symlinks inside tarball
Handle symlinks when loading a Docker tarball; also increase robustness
by normalizing the filenames in the tarball when those are used as keys
for validating metadata; with this, the loader becomes more lenient
regarding how names inside the tarball are specified; for example
<dir>/<filename> and ./<dir>/<filename> are treated equally now.
This is relevant for offline updates which as of now start to accept
Lockboxes containing Docker tarballs with symlinks.

This commit also introduces a new build target 'dockertarballloader'
which can be used for stand-alone testing the loading process.

Related-to: TOR-2426

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>