Fix power-off safety in DockerComposeSecondary
This introduces a new call that allows secondaries to perform tidy-up
operations on startup. The specific problem this fixes is the following:

1) Installation of a docker-compose secondary gets to sendFirmware() (or
   later)
2) The power goes off before Aktualizr has recorded the 'pending' update

In this case, the docker compose.tmp file will remain in place, which
causes systemd to not start them on boot, but Aktualizr won't know that
cleanup is needed (because the update isn't pending until the install
step finishes).