Do not perform Docker image pruning in some error cases
Currently we have a well-known limitation in which containers can be
pruned if they are very short lived (life measured in seconds). With
this commit we avoid having to extend that limitation to not so
short-lived containers; this is done by only pruning containers if the
application has just been stood up.

Also, to slightly improve power-cut robustness, we change the order of
pruning and temporary compose file renaming.

Finally, also rollback a change to completePendingInstall() which was
causing issues with synchronous updates.

Related-to: TOR-2972

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>