Improve robustness of CommandRunner
When launching a process, return an error rather than throw an exception
if the child process fails to launch. We can only report true/false at
the moment, so consider 'failed to start' the same as 'failed
immediately with a non-zero exit code'

When we want to abort a process, use two techniques to stop it:
 - Try killing the child with SIGINT, SIGTERM and finally
   SIGKILL.
-  Contain CommandRunner children with a process group

The latter fixes an issue where Aktualizr can't abort docker pull, which
can occur when an offline update becomes available during a normal
online update. A simple 'kill' sent to the child doesn't work because
/usr/bin/docker-compose is a wrapper that doesn't propagate the
terminate signal to its children. Running the whole thing in a process
group fixes that.