Remove targets/latest_targets distinction
This removes the distinction between 'the last director targets the
server sent us' and 'the last _non-empty_ targets the server sent us'.
This change was introduced in 2019 by https://github.com/advancedtelematic/aktualizr/pull/1186
(aka #02deeb5), we believe to deal with the old director that needed the
client to always send back a end-state manifest for any correlationid it
gave out.

Unfortunately this doesn't work with offline updates. Consider the
following steps:

1) The user sends an online update to a device
2) Aktualizr receives a non-empty targets, and installs it
3) The user plugs in a device and installs an offline update
4) Aktualizr sends the manifest to the server
5) The server recognises that the device has been updated offline, and
   starts sending empty targets to mean 'I don't know what you should be
   running, carry on as before'

Under the old behaviour, Aktualizr would ignore the empty targets and
try and install the whatever was sent in step #2.

This also removes the targets_meta field from the UpdateCheck. This
appears to be unused and doesn't have any tests. It was introduced in
2018 by https://github.com/advancedtelematic/aktualizr/pull/981, aka
Commit #c3ca9f52. We populate it by fishing the information out of the
Sqlite storage, and this introduces a few more error paths in
SotaUptaneClient.