Upgrade the minimum compiler C++17
See discussion on https://github.com/uptane/aktualizr/issues/106

Full C++17 support came in gcc 8 (although most of it was present in gcc
7 already. In terms of our likely target platforms already ship that:

* Debian oldstable shipped with gcc-10.
* The oldest supported Yocto release is Dunfell, which shipped gcc 9.5

That suggests that getting a C++17 compiler shouldn't be a huge problem
for our downstream.

Why not go further? C++20 support would require gcc 10 or 11, and while
we could go that far it seems like we should be conservative by default.