Introduce finer-grained cancellation of operations
Push the FlowControlToken (which allows aborts of in-progress operations
in the CommandQueue) further down the call stack.

This is one part of allowing offline and online updates to co-exist:
if an online update is in progress over a very slow GSM modem, then it
should be possible to plug in a USB stick, cancel the in-progress update
and start installing the new one inside a few seconds.