Move CorrelationID up a level out of Target
The Correlation ID logically belongs to a particular update that
is assigned by the server, and not to a 'target', which is one of
the things to be installed. Move the information about the current
correlation id into the director, and explictly persist it in
InvStorage.

Eventually I think we need to move torward a model where we have
an explicit representation of an 'installation job'--including a
correlation id--which persists over reboots. At the moment we infer this
from the 'pending' flag for individual ECU installations.