Tidy up RepositoryType
RepositoryType was a weird mix of a proper class type, a string and an int.
Tidy this up to use the RepositoryType class everwhere, and only convert to
a std::string when needed.

This is a particularly nice improvement in the Uptane::Expection
implementations, where we now have proper types that can be type checked.