Fix check_not_expired_targets test
When running the test "check_not_expired_targets" on a system where the
temporary directory is hosted on a different filesystem than the
directory where the tests are run, the said test fails with:

Exception: boost::filesystem::copy_file: Invalid cross-device link:...

This happens even if one sets the TMPDIR environment variable to point
to a location inside the same filesystem as where the tests are run. The
reason for this is that the test generates a temporary directory without
honoring the TMPDIR variable.

The present commit fixes the issue by creating the temporary directory
in a way that TMPDIR is honored.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>